requirements.txt
beautifulsoup4==4.4.0
Django==1.8.3
django-bootstrap3==6.1.0
django-registration-redux==1.2
factory-boy==2.5.2
fake-factory==0.5.2
gnureadline==6.3.3
html5lib==0.999999
ipython==3.2.1
Pillow==2.9.0
psycopg2==2.6.1
selenium==2.47.1
six==1.9.0
sorl-thumbnail==12.3
splinter==0.7.2
wheel==0.24.0
